previous-gen-apps/voice-kraliki/packages/ai-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-core"
version = "0.1.0"
description = "Unified LLM provider abstraction for text generation and streaming"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Platform 2026", email = "platform@example.com" }
]
keywords = ["ai", "llm", "openai", "anthropic", "gemini", "streaming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.50.0"]
gemini = ["google-generativeai>=0.8.0"]
mcp = ["mcp>=1.0.0"]
all = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "mcp>=1.0.0",
]
fastapi = ["fastapi>=0.110.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.urls]
"Homepage" = "https://github.com/m-check1B/voice-kraliki"
"Bug Tracker" = "https://github.com/m-check1B/voice-kraliki/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
